Complete 5.11.
Clean up the work to make part a and part b work with the current version of
machine by overwriting the procedures in machine which a and b require.

I have decided to not rewrite the previous commits since there were so many
changes. This will help in case something goes wrong in part a or b.